stages bien puestos y intro y fotos

+++ src/game/game.h
@@ -17,8 +17,7 @@ public:
 	static Game* instance;
 
 	std::map<int, Stage*> stages;
-	void goToStage(uint8_t stage_id);
-
+	void goToStage(int stage_id);
 
 	//window
 	SDL_Window* window;
